Gergana edits
Could you the info on what web scraping is before line 35, and add why one might want to do that, for what purposes could it be useful?

I recently found out the official Hadley R Style guide says when comments are inline with code, they should go like this:

code <- as.whatver(code)  # This is random code. - i.e. code, two spaces, #, space, text - in case you wondered why I added some spaces here and there.

At the end, perhaps put a screenshot in of what the final data frame looks like, just to kind of demonstrate the power of web scrapping? Or say some concluding remarks on what we can do now with this data frame / our newly acquired webscraping skills.

Perhaps add a bit more info on why we might be wanting to do something like that, and just highlight how this is so much better than the alternative of copying and pasting loads and loads of things in excel, where things copy across awkwardly, etc. 

Is there a way to add info on when the html pages were accessed, as a way to control for broken links/information on websites changing?

Looks really cool, I'm looking forward to going through it myself!